src/test/kotlin/com/ryanpmartz/aoc/twentytwo/Day15Test.kt
package com.ryanpmartz.aoc.twentytwo

import com.ryanpmartz.aoc.twentytwo.Day15.parsePairsFromLines
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Test

class Day15Test {


    @Test
    fun `test parsing input`() {
        val lines = listOf(
            "Sensor at x=2, y=18: closest beacon is at x=-2, y=15",
            "Sensor at x=9, y=16: closest beacon is at x=10, y=16",
            "Sensor at x=13, y=2: closest beacon is at x=15, y=3",
            "Sensor at x=12, y=14: closest beacon is at x=10, y=16",
            "Sensor at x=10, y=20: closest beacon is at x=10, y=16",
            "Sensor at x=14, y=17: closest beacon is at x=10, y=16",
            "Sensor at x=8, y=7: closest beacon is at x=2, y=10",
            "Sensor at x=2, y=0: closest beacon is at x=2, y=10",
            "Sensor at x=0, y=11: closest beacon is at x=2, y=10",
            "Sensor at x=20, y=14: closest beacon is at x=25, y=17",
            "Sensor at x=17, y=20: closest beacon is at x=21, y=22",
            "Sensor at x=16, y=7: closest beacon is at x=15, y=3",
            "Sensor at x=14, y=3: closest beacon is at x=15, y=3",
            "Sensor at x=20, y=1: closest beacon is at x=15, y=3"
        )

        val pairs = parsePairsFromLines(lines)
        assertEquals(pairs.size, 14)

        val pair = pairs.filter {
            it.sensor.x == 14 && it.sensor.y == 3
        }[0]

        assertEquals(pair.coverageRange, 1)

    }
}